remove systray app and 'runBrowser' func from skywire-visor
