+++ cmd/skywire-visor/commands/init.go
@@ -1,0 +1,104 @@
+package commands
+
+import (
+	"bytes"
+	"io"
+	"io/ioutil"
+	"net/http"
+	"os"
+	"path/filepath"
+
+	"github.com/pkg/profile"
+	"github.com/skycoin/skycoin/src/util/logging"
+
+	"github.com/skycoin/skywire/pkg/syslog"
+	"github.com/skycoin/skywire/pkg/visor/hypervisorconfig"
+	"github.com/skycoin/skywire/pkg/visor/visorconfig"
+)
+
+func initLogger(tag string, syslogAddr string) *logging.MasterLogger {
+	mLog := logging.NewMasterLogger()
+	if syslogAddr != "" {
+		hook, err := syslog.SetupHook(syslogAddr, tag)
+		if err != nil {
+			mLog.WithError(err).Error("Failed to connect to the syslog daemon.")
+		} else {
+			mLog.AddHook(hook)
+			mLog.Out = ioutil.Discard
+		}
+	}
+	return mLog
+}
+
+func initPProf(log *logging.MasterLogger, tag string, profMode string, profAddr string) (stop func()) {
+	var optFunc func(*profile.Profile)
+
+	switch profMode {
+	case "none", "":
+	case "http":
+		go func() {
+			err := http.ListenAndServe(profAddr, nil)
+			log.WithError(err).
+				WithField("mode", profMode).
+				WithField("addr", profAddr).
+				Info("Stopped serving pprof on http.")
+		}()
+	case "cpu":
+		optFunc = profile.CPUProfile
+	case "mem":
+		optFunc = profile.MemProfile
+	case "mutex":
+		optFunc = profile.MutexProfile
+	case "block":
+		optFunc = profile.BlockProfile
+	case "trace":
+		optFunc = profile.TraceProfile
+	}
+
+	if optFunc != nil {
+		stop = profile.Start(profile.ProfilePath("./logs/"+tag), optFunc).Stop
+	}
+
+	if stop == nil {
+		stop = func() {}
+	}
+	return stop
+}
+
+func initConfig(mLog *logging.MasterLogger, confPath string) *visorconfig.V1 { //nolint
+	log := mLog.PackageLogger("visor:config")
+
+	var r io.Reader
+
+	switch confPath {
+	case visorconfig.StdinName:
+		log.Info("Reading config from STDIN.")
+		r = os.Stdin
+	case "":
+		fallthrough
+	default:
+		log.Info("Reading config from file.")
+		log.WithField("filepath", confPath).Info()
+		f, err := os.ReadFile(filepath.Clean(confPath))
+		if err != nil {
+			log.WithError(err).Fatal("Failed to read config file.")
+		}
+		r = bytes.NewReader(f)
+	}
+
+	conf, compat, err := visorconfig.Parse(log, r, confPath, visorBuildInfo)
+	if err != nil {
+		log.WithError(err).Fatal("Failed to read in config.")
+	}
+	if !compat {
+		log.Fatalf("failed to start skywire - config version is incompatible")
+	}
+	if hypervisorUI {
+		config := hypervisorconfig.GenerateWorkDirConfig(false)
+		conf.Hypervisor = &config
+	}
+	if conf.Hypervisor != nil {
+		conf.Hypervisor.UIAssets = uiAssets
+	}
+	return conf
+}
