--- cmd/skywire-visor/commands/systray.go
@@ -1,52 +1,0 @@
-//go:build systray
-// +build systray
-
-package commands
-
-import (
-	"context"
-
-	"github.com/skycoin/skycoin/src/util/logging"
-	"github.com/skycoin/systray"
-
-	"github.com/skycoin/skywire/internal/gui"
-)
-
-func runApp(args ...string) {
-	l := logging.NewMasterLogger()
-	sysTrayIcon, err := gui.ReadSysTrayIcon()
-	if err != nil {
-		l.WithError(err).Fatalln("Failed to read system tray icon")
-	}
-
-	conf := initConfig(l, confPath)
-
-	go func() {
-		runVisor(conf)
-		systray.Quit()
-	}()
-
-	systray.Run(gui.GetOnGUIReady(sysTrayIcon, conf), gui.OnGUIQuit)
-
-}
-
-func setStopFunction(log *logging.MasterLogger, cancel context.CancelFunc, fn func() error) {
-	stopVisorWg.Add(1)
-	defer stopVisorWg.Done()
-
-	stopVisorFn = func() {
-		if err := fn(); err != nil {
-			log.WithError(err).Error("Visor closed with error.")
-		}
-		cancel()
-		stopVisorWg.Wait()
-	}
-
-	gui.SetStopVisorFn(func() {
-		stopVisorFn()
-	})
-}
-
-func quitSystray() {
-	systray.Quit()
-}
